+++ db/migrate/20210422200517_add_volume_label_to_digitization_book.rb
@@ -1,0 +1,6 @@
+class AddVolumeLabelToDigitizationBook < ActiveRecord::Migration[6.0]
+  def change
+    add_column :digitization_books, :volume_label, :string
+    add_rdf_column_annotation :digitization_books, :volume_label, has_predicate: ::TERMS[:rdfs].ch_label
+  end
+end